Reorganize span repository into services
